refactor(services): split SessionService and AgentService into focused services (Stream E)
Stream E.1 - SessionService Split:
- Create SessionCrudService (200 LOC) - CRUD operations
- Create SessionPresenceService (100 LOC) - User presence management
- Create SessionStreamService (300 LOC) - Event streaming/persistence
- SessionService remains as backward-compatible facade
- Add barrel file src/services/session/index.ts
- Extract shared types to src/services/session/types.ts

Stream E.2 - AgentService Split:
- Create AgentCrudService (173 LOC) - CRUD operations
- Create AgentExecutionService (506 LOC) - Agent lifecycle
- Create AgentQueueService (65 LOC) - Queue management (stub)
- AgentService remains as backward-compatible facade
- Add barrel file src/services/agent/index.ts
- Extract shared types to src/services/agent/types.ts

Design Pattern:
- Facade pattern maintains 100% backward compatibility
- Existing code using SessionService/AgentService works unchanged
- New code can import focused services for clearer dependencies

All 3002 tests passing.
