+++ src/services/agent/agent-queue.service.ts
@@ -1,0 +1,73 @@
+import type { AgentError } from '../../lib/errors/agent-errors.js';
+import type { ConcurrencyError } from '../../lib/errors/concurrency-errors.js';
+import { ConcurrencyErrors } from '../../lib/errors/concurrency-errors.js';
+import type { Result } from '../../lib/utils/result.js';
+import { err, ok } from '../../lib/utils/result.js';
+import type { Database } from '../../types/database.js';
+import type { QueuePosition, QueueStats } from './types.js';
+
+/**
+ * AgentQueueService handles queue management for agents.
+ *
+ * Responsibilities:
+ * - Queue task execution when concurrency limits are reached
+ * - Track queue positions and waiting times
+ * - Priority handling for queued tasks
+ * - Provide queue statistics
+ *
+ * Note: Queue functionality is not yet fully implemented.
+ * These methods return placeholder values indicating empty queues.
+ */
+export class AgentQueueService {
+  private readonly db: Database;
+
+  /**
+   * @param db Database instance reserved for future queue implementation.
+   *           Queue functionality will use this to persist queue state.
+   */
+  constructor(db: Database) {
+    this.db = db;
+    // Suppress unused warning - db will be used when queue is implemented
+    void this.db;
+  }
+
+  /**
+   * Queue a task for execution when agent availability permits.
+   * Currently returns QUEUE_FULL as queue functionality is not yet implemented.
+   */
+  async queueTask(
+    _projectId: string,
+    _taskId: string
+  ): Promise<Result<QueuePosition, ConcurrencyError>> {
+    return err(ConcurrencyErrors.QUEUE_FULL(0, 0));
+  }
+
+  /**
+   * Get the queue position for an agent.
+   * Currently returns null as queue functionality is not yet implemented.
+   */
+  async getQueuePosition(_agentId: string): Promise<Result<QueuePosition | null, AgentError>> {
+    // Queue functionality is not yet implemented - return null indicating not queued
+    return ok(null);
+  }
+
+  /**
+   * Get queue statistics for a project or globally.
+   * Currently returns empty stats as queue functionality is not yet implemented.
+   */
+  async getQueueStats(_projectId?: string): Promise<Result<QueueStats, never>> {
+    return ok({
+      totalQueued: 0,
+      averageCompletionMs: 0,
+      recentCompletions: 0,
+    });
+  }
+
+  /**
+   * Get all queued tasks for a project or globally.
+   * Currently returns empty array as queue functionality is not yet implemented.
+   */
+  async getQueuedTasks(_projectId?: string): Promise<Result<QueuePosition[], never>> {
+    return ok([]);
+  }
+}
