+++ src/services/session/index.ts
@@ -1,0 +1,27 @@
+/**
+ * Session services barrel file
+ *
+ * Re-exports all session-related services and types
+ */
+
+// Services
+export { SessionCrudService } from './session-crud.service.js';
+export { SessionPresenceService } from './session-presence.service.js';
+export { SessionStreamService } from './session-stream.service.js';
+
+// Types
+export type {
+  ActiveUser,
+  CreateSessionInput,
+  DurableStreamsServer,
+  GetEventsBySessionOptions,
+  HistoryOptions,
+  ListSessionsOptions,
+  ListSessionsWithFiltersOptions,
+  PresenceUpdate,
+  SessionEvent,
+  SessionEventType,
+  SessionServiceConfig,
+  SessionWithPresence,
+  SubscribeOptions,
+} from './types.js';
